formalizing effect analysis for assignments
assignment operator is different from other binary operators in that although evaluations of RHS and LHS are unordered, some partial ordering exist: a) read on RHS happens before write on LHS of the same object; 2) function calls on RHS happen before write on LHS.

Current effect analysis uses effect_context to solve the order of evaluation problem, which is not well suited for generating assignments, especially when there are embedded assignments and/or prefix/postfix ++/-- operators on RHS of assignments. A temporary solution here is keeping track a list of written variables on RHS of assignment, and adding them to the effect context when generating LHS. A better solution, we probably have to adopt later, is formalizing sequence points in effect analysis. Only with this we can extend Csmith to languages other than C/C++.