Fixes for defects reported by Pascal Cuoq.
Xuejun says: "Lhs.cpp is for fixing the first issue reported by Pascal;
FactUnion.cpp is for fixing the 2nd and 3rd issues.  I have run 100,000 test
cases for regression test."

`Lhs::make_random()': Handle certain cases in which unsigned bitfields are
undesirable.

`FactUnion::abstract_fact_for_assign()': If writing to an union field is
uncertain (due to dereference of a pointer which may points to a union
field or something else), mark the union as unreadable.