Add PowerPC little endian support and remove TARGET_CPU_* defines
The PowerPC version of read_time() is not being selected on
little endian because we are only catching the big endian cases:

 #if (TARGET_CPU_powerpc == 1 || TARGET_CPU_powerpc64 == 1)

This is the only place we use TARGET_CPU_*, and it is simpler (and
more common) to use standard c pre-processor defines:

 #if defined(__powerpc__) || defined(__powerpc64__)

This works on gcc, llvm/clang and IBM xlc. With this change we can
remove TARGET_CPU_* completely.