Fix quadratic behavior of Block::is_var_on_stack
When asking for larger blocks, csmith often goes into quadratic-seeming
behavior with big slowdowns:

```
    $ for size in 5 10 15 20; do time src/csmith --max-funcs 1 --nomain --max-block-size $size --seed 827638045 > /dev/null; done

    real	0m0.003s
    user	0m0.000s
    sys	0m0.000s

    real	0m12.372s
    user	0m12.268s
    sys	0m0.104s

    real	0m51.832s
    user	0m51.524s
    sys	0m0.308s

    real	4m1.167s
    user	4m0.568s
    sys	0m0.604s
```

The last of these has about 4600 lines of code and shouldn't take 4 minutes to
generate. Profiling showed that this was due to repeated linear searches of
lists of blocks' local variables. This patch adds a set field to blocks that
is more efficient to search.

With this the results are better:

```
    $ for size in 5 10 15 20; do time src/csmith --max-funcs 1 --nomain --max-block-size $size --seed 827638045 > /dev/null; done

    real	0m0.002s
    user	0m0.000s
    sys	0m0.000s

    real	0m9.929s
    user	0m9.828s
    sys	0m0.100s

    real	0m28.422s
    user	0m27.956s
    sys	0m0.276s

    real	1m38.668s
    user	1m38.072s
    sys	0m0.596s
```

Other such quadratic behaviors remain.