Don't test for Windows; test for `lrand48`.
In deciding whether or not to compile and link with our implementation
of `lrand48`, don't check for Windows.  Instead, test whether or not
the current platform has `lrand48`!  This is simpler and more direct,
and it seems more likely to work in cases where some Windows-based
environment (Cygwin?) might actually provide `lrand48`.