Use `srand48_deterministic` if it is available.
OpenBSD requires a special call to activate the standard, deterministic
behavior of `lrand48'.  Check for the special function and use it if it
is found.  (As far as I know, it is only available on OpenBSD.)

In addition to the principle change above, remove some old,
commented-out calls to `srand48' --- to avoid somebody uncommenting them
in the future and forgetting to call `srnd48_deterministic' instead!