Get the random seed from `arc4random_buf', if available.
Obtain the seed for the (deterministic) random-number generator by
calling `arc4random_buf', if that function is available.  It is
available on BSD and OS X in the standard library, and on Linux in
"libbsd".

If that function is not available, use the previous method: get the seed
from the timestamp counter.

This commit was inspired by a patch from the OpenBSD port of Csmith:
https://github.com/openbsd/ports/commit/8030557cd10b7f7ab6f2aac7140c50d9d09aa916