In `runtime` build directory, create symlinks to headers.
At `make`-time, in the build-tree `runtime/` directory, create symlinks
to the (non-generated) headers that reside in the source-tree `runtime/`
directory.  The generated headers were already generated into the build
tree, of course.  A `make clean` removes the symlinks.

If the build tree and the source tree are the same, then the symlinks
are not created or removed.

This change makes it somewhat easier for a developer to compile
Csmith-generated programs against a Csmith build tree that is separate
from the Csmith source tree.

This commit changes only the Automake-based build system.  The
CMake-based build system still needs to be updated.