Optimize with replacing variable vectors with variable sets.
Csmith analyzes and keeps track of list variables throughout the generation. Unfortunately the data
structure chosen, without much thought on the performance impact, was vector. This makes it inefficient
to look up or maintain a non-duplicate list. Many times we insert a variable to the vector after we are
sure the variable is not there. The complexity of insertion is O(n).

This PR switches to unordered set from vectors for maintaining variable lists in class FactPointTo, Block,
and Effect. If the performance improvement is sufficient, the next PR would replace more variable vectors
with variable sets.